doc: add more AsciiDoc cross-references
Instead of saying "see EXAMPLES below", say "see <<EXAMPLES,EXAMPLES>>
below" to make the man pages easier to navigate on the web.

The reason for using the more verbose <<EXAMPLES,EXAMPLES>>
(instead of <<EXAMPLES>>) is that if the header that `<<EXAMPLES>>`
is referring to is in an included page (for example `REMOTES` in the
`git-push` man page), then AsciiDoc will think it's a broken link even
though it isn't. So it's easier to just make all of the links use the
form with two parts.

Signed-off-by: Julia Evans <julia@jvns.ca>