config: add flag to bail when no config files exist
Add an argument for `do_git_config_sequence()` to return nonzero when
all files in the sequence are nonexistent or cannot be read for some
other reason.

When bailing, the exit code is not determined by sum of the return codes
of the underlying operations. Instead, the exit code is modified via a
single decrement. If this is undesirable, we can change it to sum the
return codes of the underlying operations instead.

The next patch changes how `git config list --global` reads the global
configuration. It uses this new flag to ensure the command continues to
bail (as expected) when both global config files are nonexistent.

Signed-off-by: Delilah Ashley Wu <delilahwu@microsoft.com>